Add a list of all PlannerInfo "roots" to PlannerGlobal.
This can serve the same purpose as glob->subplanNames, namely, to make
sure that we don't assign the same name to more than one subplan, but it
also can be used to locate a given subplan name, which is something that
glob->subplanNames cannot do.

This commit does not remove glob->subplanNames, but it ensures that
glob->allroots and glob->subplanNames have equivalent contents at the
key point in the code, namely, the start of choose_plan_name.